Simulation mode added
