+++ model/concrete_model/instantiate_params.py
@@ -1,84 +1,36 @@
-"""
-Creates the class MIMOSA:
-This is the main class. It builds a new AbstractModel
-using the chosen damage and objective modules, then reads in the
-parameter values and data (from the DataStore). With these values,
-it creates an `instance` of the AbstractModel. This is then sent to the solver.
-Finally, the export functions are called here.
-"""
-
-import os
 import numpy as np
-
-from model.common import (
-    AbstractModel,
-    TransformationFactory,
-    SolverFactory,
-    SolverManagerFactory,
-    SolverStatus,
-    value,
-    OptSolver,
-    data,
-    regional_params,
-    utils,
-    units,
-)
-from model.export import full_plot, visualise_ipopt_output, save_output
-from model.abstract_model import create_abstract_model
+from model.common import AbstractModel, units
+from model.common.data import DataStore
+from model.common.regional_params import RegionalParamStore
 
 # Util to create a None-indexed dictionary for scalar components
 # (see https://pyomo.readthedocs.io/en/stable/working_abstractmodels/data/raw_dicts.html)
 V = lambda val: {None: val}
 
 
-class MIMOSA:
-    """
-    The MIMOSA object creates an AbstractModel, makes a Concrete instance
-    out of it, solves it and saves the results.
+class InstantiatedModel:
+    def __init__(
+        self,
+        abstract_model: AbstractModel,
+        regional_param_store: RegionalParamStore,
+        data_store: DataStore,
+        quant: units.Quantity,
+    ):
 
-    Args:
-        params (dict): contains all Param values, and is based on `input/config.yaml`
+        self.abstract_model = abstract_model
+        self.regional_param_store = regional_param_store
+        self.data_store = data_store
+        self.quant = quant
 
-    Attributes:
-        params (dict)
-        regions (dict): taken from params
-        quant (Quantity): callable object used to parse and convert quantities with units
-        abstract_model (AbstractModel): the AbstractModel created using the chosen damage/objective modules
-        data_store (DataStore): object used to access regional data from the input database
-        m (ConcreteModel): concrete instance of `abstract_model`
+        # Get the non-regional params (from config.yaml) from the RegionalParamStore
+        self.params = regional_param_store.params
 
-    """
+        # First, set the data functions in the abstract model
+        self.set_data_functions()
 
-    def __init__(self, params: dict):
-        self.params = params
-        self.regions = params["regions"]
-        self.quant = units.Quantity(params)
+        self.concrete_model = self.create_instance()
 
-        self.abstract_model = self.get_abstract_model()
-        self.m = self.create_instance()
-        self.preparation()
-
-    def get_abstract_model(self) -> AbstractModel:
-        """
-        Returns:
-            AbstractModel: model corresponding to the damage/objective module combination
-        """
-        damage_module = self.params["model"]["damage module"]
-        objective_module = self.params["model"]["objective module"]
-
-        return create_abstract_model(damage_module, objective_module)
-
-    @utils.timer("Concrete model creation")
-    def create_instance(self):
-
-        # Create the regional parameter store
-        self.regional_param_store = regional_params.RegionalParamStore(self.params)
-
-        # Create the data store
-        self.data_store = data.DataStore(
-            self.params, self.quant, self.regional_param_store
-        )
-
+    def set_data_functions(self):
         # The data functions need to be changed in the abstract model
         # before initialization.
         self.abstract_model.baseline_emissions = self.data_store.data_object("baseline")
@@ -90,6 +42,12 @@ class MIMOSA:
             "carbon_intensity"
         )
         self.abstract_model.TFP = self.data_store.data_object("TFP")
+
+    def create_instance(self):
+        instance_data = self.get_param_values()
+        return self.abstract_model.create_instance(instance_data)
+
+    def get_param_values(self):
 
         damage_module = self.params["model"]["damage module"]
 
@@ -125,61 +83,7 @@ class MIMOSA:
                 )
             )
 
-        m = self.abstract_model.create_instance(instance_data)
-        return m
-
-    def preparation(self) -> None:
-
-        if len(self.regions) > 1:
-            TransformationFactory("contrib.aggregate_vars").apply_to(self.m)
-        TransformationFactory("contrib.init_vars_midpoint").apply_to(self.m)
-        TransformationFactory("contrib.detect_fixed_vars").apply_to(self.m)
-        if len(self.regions) > 1:
-            TransformationFactory("contrib.propagate_fixed_vars").apply_to(self.m)
-
-    @utils.timer("Model solve")
-    def solve(
-        self,
-        verbose=False,
-        halt_on_ampl_error="no",
-        use_neos=False,
-        neos_email=None,
-        ipopt_output_file=None,
-    ) -> None:
-        if use_neos:
-            os.environ["NEOS_EMAIL"] = neos_email
-            solver_manager = SolverManagerFactory("neos")
-            solver = "conopt"  # 'ipopt'
-            results = solver_manager.solve(self.m, opt=solver)
-        else:
-            opt: OptSolver = SolverFactory("ipopt")
-            opt.options["halt_on_ampl_error"] = halt_on_ampl_error
-            opt.options["output_file"] = ipopt_output_file
-            results = opt.solve(self.m, tee=verbose, symbolic_solver_labels=True)
-            if ipopt_output_file is not None:
-                visualise_ipopt_output(ipopt_output_file)
-
-        # Restore aggregated variables
-        if len(self.regions) > 1:
-            TransformationFactory("contrib.aggregate_vars").update_variables(self.m)
-
-        if results.solver.status != SolverStatus.ok:
-            print(
-                "Status: {}, termination condition: {}".format(
-                    results.solver.status, results.solver.termination_condition
-                )
-            )
-            if results.solver.status != SolverStatus.warning:
-                raise Exception("Solver did not exit with status OK")
-
-        print("Final NPV:", value(self.m.NPV[self.m.tf]))
-
-    @utils.timer("Plotting results")
-    def plot(self, filename="result", **kwargs):
-        full_plot(self.m, filename, **kwargs)
-
-    def save(self, experiment=None, **kwargs):
-        save_output(self.params, self.m, experiment, **kwargs)
+        return instance_data
 
     ########################
     #
